+++ src/main/java/org/academiadecodigo/simplegraphics/graphics/Fillable.java
@@ -1,4 +1,4 @@
-package com.codeforall.simplegraphics.graphics;
+package org.academiadecodigo.simplegraphics.graphics;
 
 /**
  * Methods for filling a shape
