+++ src/main/java/org/academiadecodigo/simplegraphics/graphics/Rectangle.java
@@ -1,22 +1,25 @@
-package com.codeforall.simplegraphics.graphics;
+package org.academiadecodigo.simplegraphics.graphics;
 
 import java.awt.*;
-import java.awt.geom.Ellipse2D;
+import java.awt.geom.Rectangle2D;
 
 /**
- * Represents an ellipse shape that can be drawn, filled, moved, and resized.
- * The ellipse is defined by its bounding box, which determines its x/y position, width, and height.
- * This class integrates with the {@link Canvas} to visually render itself.
+ * The {@code Rectangle} class represents a drawable and optionally filled rectangle
+ * that can be displayed on a canvas. This class allows manipulation of the rectangle's
+ * size, position, and color, and provides basic drawing and transformation capabilities.
  *
  * Implements:
- * <ul>
- *   <li>{@link Shape} – for rendering and bounding box info</li>
- *   <li>{@link Colorable} – for setting its color</li>
- *   <li>{@link Fillable} – for supporting fill operations</li>
- *   <li>{@link Movable} – for translation and scaling</li>
- * </ul>
+ *  <ul>
+ *      <li>{@link Shape} – for rendering and bounding box info</li>
+ *      <li>{@link Colorable} – for setting its color</li>
+ *      <li>{@link Fillable} – for supporting fill operations</li>
+ *      <li>{@link Movable} – for translation and scaling</li>
+ *  </ul>
+ *
+ * <strong>Note:</strong> This class is distinct from {@code java.awt.Rectangle}
+ * and is intended for use within the simplegraphics library.
  */
-public class Ellipse implements Shape, Colorable, Fillable, Movable {
+public class Rectangle implements Shape, Colorable, Fillable, Movable {
 
     private Color color = Color.BLACK;
     private boolean filled = false;
@@ -26,14 +29,24 @@ public class Ellipse implements Shape, Colorable, Fillable, Movable {
     private double height;
 
     /**
-     * Constructs an ellipse.
+     * Constructs an empty rectangle.
+     */
+    public Rectangle() {
+        x = 0;
+        y = 0;
+        width = 0;
+        height = 0;
+    }
+
+    /**
+     * Constructs a rectangle.
      *
      * @param x      the leftmost x-coordinate
      * @param y      the topmost y-coordinate
-     * @param width  the width of the bounding box
-     * @param height the height of the bounding box
+     * @param width  the width
+     * @param height the height
      */
-    public Ellipse(double x, double y, double width, double height) {
+    public Rectangle(double x, double y, double width, double height) {
         this.x = x;
         this.y = y;
         this.width = width;
@@ -41,7 +54,7 @@ public class Ellipse implements Shape, Colorable, Fillable, Movable {
     }
 
     /**
-     * Gets the leftmost x-position of this ellipse.
+     * Gets the leftmost x-position of this rectangle.
      *
      * @return the leftmost x-position
      */
@@ -51,7 +64,7 @@ public class Ellipse implements Shape, Colorable, Fillable, Movable {
     }
 
     /**
-     * Gets the topmost y-position of this ellipse.
+     * Gets the topmost y-position of this rectangle.
      *
      * @return the topmost y-position
      */
@@ -60,9 +73,8 @@ public class Ellipse implements Shape, Colorable, Fillable, Movable {
         return (int) Math.round(y);
     }
 
-
     /**
-     * Gets the width of the bounding box.
+     * Gets the width of this rectangle.
      *
      * @return the width
      */
@@ -72,7 +84,7 @@ public class Ellipse implements Shape, Colorable, Fillable, Movable {
     }
 
     /**
-     * Gets the height of the bounding box.
+     * Gets the height of this rectangle.
      *
      * @return the height
      */
@@ -82,7 +94,7 @@ public class Ellipse implements Shape, Colorable, Fillable, Movable {
     }
 
     /**
-     * Moves this ellipse by a given amount.
+     * Moves this rectangle by a given amount.
      *
      * @param dx the amount by which to move in x-direction
      * @param dy the amount by which to move in y-direction
@@ -95,12 +107,11 @@ public class Ellipse implements Shape, Colorable, Fillable, Movable {
     }
 
     /**
-     * Resizes this ellipse both horizontally and vertically.
+     * Resizes this rectangle both horizontally and vertically.
      *
      * @param dw the amount by which to resize the width on each side
      * @param dh the amount by which to resize the height on each side
      */
-    @Override
     public void grow(double dw, double dh) {
         width += 2 * dw;
         height += 2 * dh;
@@ -110,7 +121,7 @@ public class Ellipse implements Shape, Colorable, Fillable, Movable {
     }
 
     /**
-     * Sets the color of this ellipse.
+     * Sets the color of this rectangle.
      *
      * @param newColor the new color
      */
@@ -121,7 +132,7 @@ public class Ellipse implements Shape, Colorable, Fillable, Movable {
     }
 
     /**
-     * Draws this ellipse on the canvas.
+     * Draws this rectangle.
      */
     @Override
     public void draw() {
@@ -130,7 +141,7 @@ public class Ellipse implements Shape, Colorable, Fillable, Movable {
     }
 
     /**
-     * Deletes this ellipse from the canvas
+     * Deletes this rectangle
      */
     @Override
     public void delete() {
@@ -138,7 +149,7 @@ public class Ellipse implements Shape, Colorable, Fillable, Movable {
     }
 
     /**
-     * Fills this ellipse.
+     * Fills this rectangle.
      */
     @Override
     public void fill() {
@@ -147,31 +158,30 @@ public class Ellipse implements Shape, Colorable, Fillable, Movable {
     }
 
     /**
-     * Returns a string representation of this ellipse.
+     * Returns a string representation of the rectangle, including both endpoints.
      *
-     * @return a string describing the ellipse's position and dimensions
+     * @return a string describing the rectangle's coordinates
      */
     @Override
     public String toString() {
-        return "Ellipse[x=" + getX() + ",y=" + getY() + ",width=" + getWidth() + ",height=" + getHeight() + "]";
+        return "Rectangle[x=" + getX() + ",y=" + getY() + ",width=" + getWidth() + ",height=" + getHeight() + "]";
     }
 
     /**
-     * Paints this ellipse using the given Graphics2D context.
-     * This is called by the canvas to render the shape.
+     * Paints the rectangle using the specified Graphics2D context.
      *
-     * @param g2D the Graphics2D context to draw with
+     * @param g2D the graphics context to paint with
      */
     @Override
     public void paintShape(Graphics2D g2D) {
-        Ellipse2D.Double ellipse = new Ellipse2D.Double(getX(), getY(), getWidth(), getHeight());
+        Rectangle2D.Double rect = new Rectangle2D.Double(getX(), getY(), getWidth(), getHeight());
         g2D.setColor(new java.awt.Color((int) color.getRed(), (int) color.getGreen(), (int) color.getBlue()));
 
         if (filled) {
-            g2D.fill(ellipse);
+            g2D.fill(rect);
 
         } else {
-            g2D.draw(ellipse);
+            g2D.draw(rect);
         }
     }
 }
