+++ src/main/java/org/academiadecodigo/simplegraphics/graphics/Shape.java
@@ -1,4 +1,4 @@
-package com.codeforall.simplegraphics.graphics;
+package org.academiadecodigo.simplegraphics.graphics;
 
 import java.awt.*;
