+++ src/main/java/org/academiadecodigo/simplegraphics/graphics/Text.java
@@ -1,4 +1,4 @@
-package com.codeforall.simplegraphics.graphics;
+package org.academiadecodigo.simplegraphics.graphics;
 
 import javax.swing.*;
 import java.awt.*;
