+++ src/main/java/org/academiadecodigo/simplegraphics/keyboard/KeyboardEventType.java
@@ -1,4 +1,4 @@
-package com.codeforall.simplegraphics.keyboard;
+package org.academiadecodigo.simplegraphics.keyboard;
 
 /**
  * The type of events supported by the Keyboard
